WIP searchIndexing in backend
work on having comparison arrays created for each piece of JSON data in the backend. these will be passed along to the front-end to aid with searching.
also reformatted the JSON data, changed preview.js a bit to work nicely with those changes.
added the ability to click out of the showcase component.